+++ web/src/App.tsx
@@ -20,6 +20,7 @@ import { Project } from "./routes/Project";
 import { Report } from "./routes/Report";
 import { Tracker } from "./routes/Tracker";
 import { Scheduling } from "./routes/Scheduling";
+import { Disciplines } from "./routes/Disciplines";
 import { Showcase } from "./routes/Showcase";
 import { NotFound, Todos } from "./routes/placeholders";
 
@@ -60,6 +61,7 @@ export function App() {
         <Route path="todos" element={<Todos />} />
         <Route path="tracker/:name" element={<Tracker />} />
         <Route path="scheduling" element={<Scheduling />} />
+        <Route path="disciplines" element={<Disciplines />} />
         <Route path="*" element={<NotFound />} />
       </Route>
     </Routes>